+++ variables.tf
@@ -228,6 +228,75 @@ variable "enable_azure_bastion" {
   description = "Specify if Azure Bastion should be provisioned. See https://azure.microsoft.com/en-us/services/azure-bastion for more information."
 }
 
+variable "enable_hybrid_benefit_server_licenses" {
+  default     = false
+  type        = bool
+  description = "Enable Azure Hybrid Benefit to use your on-premises Windows Server licenses and reduce cost. See https://docs.microsoft.com/en-us/azure/virtual-machines/windows/hybrid-use-benefit-licensing for more information."
+}
+
+variable "vm_dc_size" {
+  default     = "Standard_B2s"
+  description = "Size of the DC VM."
+}
+
+variable "vm_dc_storage_account_type" {
+  default     = "StandardSSD_LRS"
+  description = "Type of storage for the managed disks. Visit 'https://docs.microsoft.com/en-us/rest/api/compute/disks/list#diskstorageaccounttypes' for more information."
+  validation {
+    condition = contains([
+      "Standard_LRS",
+      "StandardSSD_LRS",
+      "Premium_LRS",
+      "Premium_ZRS",
+      "StandardSSD_ZRS",
+      "UltraSSD_LRS"
+    ], var.vm_dc_storage_account_type)
+    error_message = "Invalid storage account type value."
+  }
+}
+
+variable "vm_sql_size" {
+  default     = "Standard_B2ms"
+  description = "Size of the SQL VM."
+}
+
+variable "vm_sql_storage_account_type" {
+  default     = "StandardSSD_LRS"
+  description = "Type of storage for the managed disks. Visit 'https://docs.microsoft.com/en-us/rest/api/compute/disks/list#diskstorageaccounttypes' for more information."
+  validation {
+    condition = contains([
+      "Standard_LRS",
+      "StandardSSD_LRS",
+      "Premium_LRS",
+      "Premium_ZRS",
+      "StandardSSD_ZRS",
+      "UltraSSD_LRS"
+    ], var.vm_sql_storage_account_type)
+    error_message = "Invalid storage account type value."
+  }
+}
+
+variable "vm_sp_size" {
+  default     = "Standard_B4ms"
+  description = "Size of the SP VM."
+}
+
+variable "vm_sp_storage_account_type" {
+  default     = "StandardSSD_LRS"
+  description = "Type of storage for the managed disks. Visit 'https://docs.microsoft.com/en-us/rest/api/compute/disks/list#diskstorageaccounttypes' for more information."
+  validation {
+    condition = contains([
+      "Standard_LRS",
+      "StandardSSD_LRS",
+      "Premium_LRS",
+      "Premium_ZRS",
+      "StandardSSD_ZRS",
+      "UltraSSD_LRS"
+    ], var.vm_sp_storage_account_type)
+    error_message = "Invalid storage account type value."
+  }
+}
+
 variable "_artifactsLocation" {
   default = "https://github.com/Azure/azure-quickstart-templates/raw/master/application-workloads/sharepoint/sharepoint-adfs/"
 }